Negotiate extmaps
This patch add extmaps negotiation

Every extmap has a numeric value the must be negotiated and be the same
between offerer and answer. It's the answerer that should accept the
extmap value offered by the offerer.

Additionally these value must be the same for every media section since,
for example, the mid extension extmap is used for getting the related
mid section of the rtp packet.

Before this patch pion was using only, when required, the transportcc
extension.  Now we also add the mid and streamId extension (required for
simulcast) We have an initial extmaps list with predefined values,
they'll be used if we're offering, if we receive an offer where the
extmap value is different we accept the offered value.